requirements.txt
clip==0.2.0
einops==0.4.1
matplotlib==3.5.1
numpy==1.19.5
opencv_python==4.5.5.64
Pillow==9.0.1
Pillow==10.2.0
torch==1.8.2+cu111
torchvision==0.9.2+cu111
tqdm==4.63.0
